Fix support for using a private CA in the ingresses (#437)
Atm the support for using a custom certificate or a private CA in the dex ingresses is partially broken. Since the cert-manager.io/cluster-issuer annotation was always set, cert-manager will always try to manage the certificate Secret, and will "fight" any other component trying to manage it.

With this cert-manager can be disabled altogether so the `dex-tls` Secret can be managed externally.

I've also made the config options more on-par with the rest of the apps as per https://github.com/giantswarm/shared-configs/pull/404

Note that when this is rolled out, cert-manager will re-issue the Ingress certificates since the Secret names have changed